rename keygenerator file
